+++ crates/tycho-client-py/python/tycho_indexer_client/dto.py
@@ -48,6 +48,7 @@ class Chain(str, Enum):
     polygon = "polygon"
     plasma = "plasma"
     robinhood = "robinhood"
+    arc = "arc"
 
 
 class CustomChain(BaseModel):